requirements.txt
aniso8601==9.0.1
click==8.0.0
ecdsa==0.14.1
Flask==2.0.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pyasn1==0.4.8
python-jose==3.2.0
pytz==2021.1
rsa==4.7.2
six==1.16.0
Werkzeug==2.0.0
